Base interface & keyboard

+++ emulator/io/keyboard/Keyboard.cpp
@@ -26,7 +26,7 @@ AddressSpace Keyboard::getSpace() const
     };
 }
 
-void Keyboard::keyevent(unsigned short code, bool isPressed)
+void Keyboard::keyevent(int code, bool isPressed)
 {
     for (int i = 0; i < 8; i++)
     for (int j = 0; j < 8; j++)
